WRO-10782: Update pattern-ls2request samples to use Redux Toolkit (#181)
* pattern-ls2request update to use redux toolkit

* updated ui/pattern-ls2request  to use redux toolkit

* Minor fix

* Moved sandstone sample to redux-toolkit.

* Lint fix

* Review fixes

* Update ui/pattern-ls2request/src/store/store.js


* rename the store file

Enact-DCO-1.0-Signed-off-by: Mikyung Kim (mikyung27.kim@lge.com)
