+++ sandstone/pattern-ls2request-camera/src/components/CameraView.js
@@ -5,9 +5,9 @@ import VirtualList from '@enact/sandstone/VirtualList';
 import {Column} from '@enact/ui/Layout';
 import ri from '@enact/ui/resolution';
 import {useEffect, useRef} from 'react';
-import {useSelector, useDispatch} from 'react-redux';
+import {useDispatch, useSelector} from 'react-redux';
 
-import {closeCamera, getCameraIds, startCamera} from '../actions';
+import {closeCamera, getCameraIds, startCamera} from '../store';
 
 const CameraView = () => {
 	const videoRef = useRef(null);
