--- sandstone/pattern-ls2request-camera/src/reducers/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "reducers.js"
-}
