+++ ui/pattern-ls2request/package.json
@@ -34,12 +34,12 @@
     "@enact/spotlight": "^4.5.0",
     "@enact/ui": "^4.5.0",
     "@enact/webos": "^4.5.0",
+    "@reduxjs/toolkit": "^1.8.4",
     "ilib": "^14.14.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^7.2.0",
-    "redux": "^4.2.0",
-    "redux-thunk": "^2.4.1"
+    "redux": "^4.2.0"
   }
 }
