+++ ui/pattern-ls2request/src/main.js
@@ -1,10 +1,7 @@
 import {Provider} from 'react-redux';
 
 import App from './App';
-import configureStore from './store';
-
-// set default launch path
-const store = configureStore();
+import store from './store';
 
 let appElement = () => (
 	<Provider store={store}>
